requirements.txt
psycopg2-binary==2.9.9
django-environ==0.11.2
django==4.2.13 #5.0.6
djoser==2.2.2
djangorestframework==3.15.1
django-debug-toolbar==4.4.2
psycopg2-binary==2.9.9
django-environ==0.11.2
python-dateutil==2.9.0
gunicorn==22.0.0
boto3==1.34.145
django-storages==1.14.2
fcm-django==2.2.1
python-dotenv==1.0.1
pytz==2023.3.post1
celery==5.4.0
django_celery_beat==2.7.0
redis==5.2.0
celery-singleton==0.3.1
google-auth-oauthlib==1.2.1
django-cors-headers==4.6.0
twilio==9.4.3
Pillow==11.1.0
django-notifications-hq==1.8.3
phonenumbers==8.12.31